fs-verity: add support for extension items
Allow storing variable-length metadata items, or "extensions", following
the fixed-size portion of the fs-verity footer.

Extensions can be either authenticated or unauthenticated.
Authenticated extensions are included in the file measurement, while
unauthenticated extensions are not.  By default extensions are
authenticated; unauthenticated extensions must be explicitly marked as
such.  Unauthenticated extensions will be needed for metadata involved
in establishing the root of trust itself, like a PKCS#7 signature.
Authenticated extensions are stored first, directly following the
fixed-size portion of the fs-verity footer; unauthenticated extensions
come afterwards.

No extensions are implemented by this patch itself, but we reserve type
codes for a salt (for variable-length salt support); a PKCS#7 signature;
and the elide and patch support required by the Android APK files.

Signed-off-by: Eric Biggers <ebiggers@google.com>

Bug: 30972906
Bug: 67380979
Change-Id: Ib0727af85513aff3700a8e5693af9f9743c10bf0