fsverity fixup
After freeing i_verity_info, it's necessary to set it NULL since the
the inode structure doesn't get cleared when it gets allocated.
That's why it's up to functions like inode_init_always() to reset
various fields to be zero.  For i_verity_info, we should do that in
fsverity_cleanup_inode().

Signed-off-by: Theodore Ts'o <tytso@mit.edu>

Bug: 30972906
Bug: 67380979
Change-Id: Ifb8f58383a94d7770063e106a2f233234ce4d5f3
Signed-off-by: Eric Biggers <ebiggers@google.com>